Dynamic type
Summary:
Introduce type `dynamic` into the formalization.
There are no additions to the subtyping rules - `dynamic` just behaves as an abstract type.
Introduce notion of *implicit coercion*, to begin with just allowing any type to be implicitly coerced to `dynamic`, and defining a notion of
an expression being coercable to a type, extended to sequences of expressions.

Next diff will modify function application and other rules to support implicit coercion.

Reviewed By: manzyuk

Differential Revision: D8677495

fbshipit-source-id: cb07e028802fcd484b7997e61733669c263640f1