try to fix a crash
Summary:
trunk is crashing upon failures to advise out the unused parts of the stack, which contains hugetlb pages.  Previously we were using s_firstSlab as an indicator that the stack has a part on huge pages.  However, s_firstSlab is no longer populated (as it cannot really be used and is being deleted in D8777364), and thus the madvise failure.
This diff tries to temporarily stop the crash by avoiding doing madvise on 2M hugetlb pages (even when no hugetlb pages are used).  Will introduce proper thread local variables to better describe stack layout and huge page usage.

Differential Revision: D8780308

fbshipit-source-id: 06d6ab64234bd5d754b68fabbe188d0362c86530