Better heuristics for context_sensitive_return_type
Summary:
hhbbc has code to figure out the return type of a function by
substituting in the actual arguments, and interpreting it.

Turning that on everywhere more than doubles the hhbbc runtime,
without providing much benefit.

This adds a heuristic which enables it for functions that return one
of their parameters.

Reviewed By: ricklavoie

Differential Revision: D8762299

fbshipit-source-id: ad62ab703c09e55c0aa7966a0be2a37b24fe9e2f