(More) sound typing of Shapes::idx
Summary:
The result of accessing a shape field of a nullable type using `Shapes::idx()` should always be nullable, even if a non-nullable default is provided (the default value is only used if the field is missing from the shape but is not used if it is present and is set to `null`).  We were failing to detect that the result should be nullable when the shape had an unresolved or abstract type (e.g., a generic type parameter constrained by a shape, or a newtype with a shape upper bound).  This diff fixes this problem.

I've replaced the ad hoc checks of whether a shape field is nullable/optional done by case analysis on the shape with a more general approach using subtyping.

I'm suggesting to kill off the `shape_idx_access_unknown_field` lint warning because it doesn't make sense to me. I've moved `shape_idx_access_required_field` to `shape_field_check.ml`.

Reviewed By: nbenton

Differential Revision: D8767688

fbshipit-source-id: c01104bf0b62fe6cf16c1ee84c52065e777bdb98