Lex trailing trivia for self-closing XHP tags in expression contexts
Summary:
Previously, we never lexed trailing trivia for XHPOpen nodes. This is because we don't want to lex trailing trivia in an XHP body context--if we encounter an XHPBody token, any preceding trivia (which would normally be lexed as trailing) should be considered leading trivia for that token, since it is semantically significant.

In this example:

```
<p>
  body
</p>
```

We would not lex the newline following `<p>` as trailing trivia (as we normally would for most tokens), since we want the newline and two spaces to be associated with the `body` token, since whitespace is semantically significant in XHP.

However, this is the wrong behavior for self-closing tags in an expression context. For example:

```
$cond
  ? <p />
  : null;
```

We would not lex the newline following `<p />`, leaving it to be leading trivia of the `:` token.

This caused hackfmt to be confused. Newlines in leading trivia (outside of an XHP body) mean blank lines in the original source. This caused hackfmt to insert blank lines after self-closing XHP tags.

This change ensures that we do lex trailing trivia for the SlashGreaterThan token of a self-closing XHP element, when in an expression context.

Reviewed By: pittsw, arxanas

Differential Revision: D8795961

fbshipit-source-id: 53bfb5b4b9b6c714d345474891da17f3afa45398