Replace sorted maps with f14 maps in stream-wrapper-registry
Summary:
The one place in the code we iterated over the std::map or req::map
of wrappers was appending them both to an array, so the result array
was not sorted. The examples in the PHP docs for listing the wrappers
also show an unsorted list.

Assuming its okay for these lists to have an upredictable order, this
reworks the code to use hashtables instead of an expensive tree map.

If the unstable order is an issue, it may yet still be just as good
(or better) to sort the result array, than to maintain the internal
tables in sorted order.

Reviewed By: markw65

Differential Revision: D8471558

fbshipit-source-id: 91026500530930100b1c9072e5afc8b708c60980