Improve lowering errors
Summary: This makes hh_single_type_check not assert failure when it encounters a lowering error, but instead give the parsing error. It also stops logging lowering errors caused by parsing errors. When we encounter a file with parsing errors, we should try to lower the AST, but failing to lower is understandable if the parsing errors cause the code to not be well structured.

Reviewed By: kmeht

Differential Revision: D8789905

fbshipit-source-id: 740cc95cb63052ce4bb512996e8a3b1a7d8d3295