fix hugestack accounting
Summary:
Try to fix a stack accounting bug introduced in the previous diff.

When allocating stack for a worker thread, we keep track of the allocation size, so that when the thread exits, we can `munmap` the stack.
The bug was that if we only use a part of a hugetlb page, we should still account for the entire page, because we cannot unmap a partial hugetlb page.

Reviewed By: markw65

Differential Revision: D8830547

fbshipit-source-id: c891b96c2868cca0ca063b77bf7d00c7027b6af9