Convert rest of typechecker tests to use HH_FLAGS
Summary:
This diff converts the rest of the typechecker files to use HH_FLAGS.

The only real change I've made is to make a null coalesce change work, and should only affect the typechecker. Essentially, we weren't checking that function call expressions were valid lvalues.

In the future, we need to change this check of lvalues from a negative to a positive test: only certain things can be lvalues. But I don't have time to do that in this diff.

Reviewed By: kmeht

Differential Revision: D8774858

fbshipit-source-id: c74a02d1b0da345e7251360dfc410ed91827c20e