Inject __RxOfScope into list of attributes for async blocks
Summary:
Async blocks are lowered to call to lambda expression reactivity of which should be the same as reactivity of containing function. There is already a machinery for do this - inject `__RxOfScope` to enable it.
Allow async methods with mutable `$this` - we already lifted the restriction for regular mutable parameters so receiver should be treated in the same way.

Reviewed By: DavidSnider

Differential Revision: D8883147

fbshipit-source-id: 3bcf784e264cd2ed37adb9391f47414ff74d9f61