Implement LSB memoization with attribute <<__MemoizeLSB>>
Summary:
Adds attribute <<__MemoizeLSB>> (static methods only) which makes
the memoization cache specific to the LSB class.

This is functionally equivalent to __Memoize combined with static::class
as a hidden cache key, but the implementation attempts to achieve greater
performance by using an array of cached handles in the Class object.

Reviewed By: ricklavoie

Differential Revision: D8850452

fbshipit-source-id: 1f08c48a2a698f12a8b64f482009494e18a15765