ignore hack keywords in php files
Summary:
Hack keywords in PHP files should be understood as simple identifiers. For example:
```
class inout {}
```
The simplest way to do that is to have the lexer aware of what it is lexing: PHP or Hack. This is done using a - for now very small - `Lexer_env` module.
The generated `Full_fidelity_token_kind` does the lexing of keywords in function `from_string`. We change the generator so that this function now accept an `is_hack` argument. We also change the `Full_fidelity_schema` so that developers now have to specify whether a given keyword is "hack_only" or "php_and_hack".

Reviewed By: jamesjwu

Differential Revision: D8735588

fbshipit-source-id: 06474fd994885dce3f1ea9fed695bf5ba7bfac33