Script to find cycles in #includes
Summary:
There are probably better and more robust ways to do this with clang,
but this was quick and easy to write, and gets the job done in all the cases
I've tried so far. It also has an option to print the header dependency graph
as a dot file.

Reviewed By: paulbiss

Differential Revision: D8305540

fbshipit-source-id: a1a665e5db69324005c5e3eb281b52576e0c7e61