AKExists optimization for {Imm,}Vector,kPackedArray
Summary:
for `array_key_exists($key, $vector_or_packed_array)` to be true:
- the key has to be an integer
- the key has to be greater or equal to 0
- the key needs to be less than the size of the vector
- the vector needs to not be empty (covered by 0<=key<size)

we can test all of those things without calling `ak_exist_int_obj()` and w/o branches.

if the integer is a constant we can do the >=zero-check manually and just test that the key is less than the count.

  if (array_key_exists(SomeEnum::FOO, $v)) {

Reviewed By: jano

Differential Revision: D2349806

fbshipit-source-id: 6cc3b1795fd13fceefb160b08a97e4def408cc07