Require functions with implicit or expressionless returns to have return type void
Summary:
Disallow returning values from functions that have implicit or expressionless (i.e., of the form `return;`) returns.  In other words, mixing `return <expr>;` with `return;` or an implicit return in the same function should not be allowed.

The way this is implemented in this diff is by requiring the return type of a function containing an implicit return or `return;` to unify with `void`.  As a consequence, function definitions like
```
function foo(): mixed {}
```
are rejected and one needs to explicitly return `null` in such cases.

Reviewed By: andrewjkennedy

Differential Revision: D8835116

fbshipit-source-id: a72cacc62e43f924d7bdf3b313f642ff30c01a83