Consolidate namespace aliasing between HackC and Hack
Summary:
Now that we have unified the behavior of namespace aliasing between the typechecker and the runtime, we can reuse code as well! This is almost a complete revert of D7216602, but maintains the same functionality guaranteed by tests, including a new test to document the behavior that namespace aliasing works even in the namespace definition.

We did make one behavioral change: `use Foo` does override namespace aliases, when previously it didn't. We plan to ban this syntax in Hack in the near future.

Because the Namespaces module uses the parser options for the namespace map (and HackC doesn't use parser options at all), I instead placed the alias map on the NS env. Currently, it's inserted in Hhbc_id right before elaborating, but in the future we should thread this to wherever namespace envs are created (quite a few places) and make sure the alias map is loaded then.

Reviewed By: oulgen

Differential Revision: D8804523

fbshipit-source-id: 2d8efa112a1504d0b50d1eda68ddf04c6983e3c2