Fix hh_server naming crash
Summary:
We crash when exposed to this code in incremental mode:

```
<?hh

function foo : void {
  try {
 } catch(e) { // no dollar sign!
  }
}
```

This diff fixes the index out of bounds by bounds checking the string first.

There are a few more instances of .[0] in naming, but none of them seem to trigger crashes, so I won't be editing them for now.

Reviewed By: dabek

Differential Revision: D8921446

fbshipit-source-id: f286d962912163aff134d00f1b06b01d33438ccf