Array append typing
Summary:
Typing rules for array append are surprisingly involved.  Here are a few examples to consider.

  * `$vs[] = 'foo';`, where `$vs : vec<int>` should be allowed and the type of `$vs` after the assignment should be `vec<int | string>`.

  * `$vs[] = 'foo';`, where `$vs : Vector<int>` should be rejected.

  * `$obj->vs[] = 'foo';`, where `$obj : C` and the class `C` contains a property `vs` of type `vec<int>`, i.e., `class C { public vec<int> $vs; ... }`, should be rejected.  On the other hand, if the property `vs` has the type `vec<arraykey>`, the assignment should be allowed.  The type of `$obj` doesn't change as a result of the assignment.

  * `$s['foo'][] = 'bar';`, where `$s : shape('foo' => vec<int>)` should be allowed and the type of `$s` after the assignment should be `shape('foo' => vec<int | string>)`.

  *  `$d['foo']->vs[] = 'bar';`, where `$d : dict<string, C>`, where `C` is as above should be rejected if the type of `vs` is `vec<int>` and should be allowed the type of `vs` is `vec<arraykey>`.  The type of `$d` is not changed by the assignment.

  * `$m['foo'][] = 'bar';`, where `$m : Map<string, vec<int>>` should be rejected.

Informally, appending to a collection with value semantics should allow the type of the collection to grow unless it is part of an object with reference semantics (including another collection).  Appending to a collection with reference semantics should be allowed only if the type of the appended value is compatible (i.e., a subtype of) the values in the collection.

In this diff, I'm trying to flesh out rules that would capture the above desiderata.

Reviewed By: CatherineGasnier

Differential Revision: D8723456

fbshipit-source-id: fb7aca0830e20316a58364b8cf881214b4506dfb