Move redundant nullsafe property access check out of typing.ml
Summary: We have a check in `typing.ml` if `?->`, nullsafe property access operator, is used on a value that is guaranteed to not be null.  This check better belongs into a TAST check (and shouldn't be a type error anyway because a value of type `C` can be viewed as a value of type `?C` by subsumption and hence it should be possible to use `?->` on it).

Reviewed By: andrewjkennedy

Differential Revision: D8893174

fbshipit-source-id: ff5d65537c3cb0ae40c3b270e750e1ad39163082