Remove AKgeneric, just use Tgeneric
Summary:
Generic parameters in the decl phase are represented using the `Tgeneric` constructor, but are transformed into the `AKgeneric` special case of `Tabstract` during localization. This is redundant, as the `as` constraint on a generic parameter is no longer stored inline (as the second argument to `Tabstract`), *except* in certain limited situations such as non-null refinement.

Therefore we can kill of `AKgeneric` and just use `Tgeneric` even after localization, relying on the type parameter environment for bounds.

Reviewed By: manzyuk

Differential Revision: D8875048

fbshipit-source-id: df944ca32769ddffcce5763ef972cc4ba08501f9