SharedMem.check_should_exit
Summary:
`SharedMem.stop_workers` sets a bit. When a worker tries to read or write the shared memory, it will check the bit and throw an exception if that bit is set. This is a clever way of stopping the workers, assuming they routinely access shared memory.

In Flow, I have some code which doesn't access the shared memory for a while. But I still want the worker to stop when this bit is set. So I'm exposing a method so that OCaml code can check the bit.

Reviewed By: jbrown215, gregory-nisbet

Differential Revision: D8789878

fbshipit-source-id: 3478b69b82d1a021e17390f75cc8ee96a3028c50