Refactor Verifier to use UnitEmitters
Summary: We want to start running the verifier on all builds of hhvm and using it to std::abort repo-authoritative builds where either the input or the output fail to verify, and to convert unverifiable units into fatals in sandbox mode. To do this we will need to be able to operate on UnitEmitter as we will not have Units available yet.

Reviewed By: alexeyt

Differential Revision: D8599506

fbshipit-source-id: 2ecd24b941e8cab373614fcfb299471ace187415