Add position tracking to class IDs.
Summary: In the NAST and TAST, we only track positions on expressions, not on class IDs.  This works fine in most cases, but since we don't have positional data for class IDs the symbol positions get messed up for constructors, and the only way to fix it is to track class ID positions.

Differential Revision: D8335742

fbshipit-source-id: e76f0c34b510a077006b6abb8b510a43251e33b2