`require extends` Find-References (and Rename) method support
Summary:
Previously, we were unable to find the references for a method of an interface that `require extended` from the parent class/interface.

Now, we access the class name that the method originated from and pass that down as the class name we find child classes for. Then, as we gather the child classes that could extend the parent class, we include the child classes that `require extend` the parent.

This also allowed me to successfully rename `toString` into `toStringWithTrivia` in `www` on all typed calls.

Reviewed By: arxanas

Differential Revision: D8908239

fbshipit-source-id: 6c50e91f38a65d9a236b7c10a59149fef2674647