Analyze Idx, ArrayIdx, and AKExists in HHBBC
Summary:
HHBBC wasn't making any effort to analyze Idx, ArrayIdx, or AKExists at all, so
do so now.

The existing array_elem, vec_elem, dict_elem, and keyset_elem functions can
mainly be used for this, with a few tweaks. The main difficulty is that these
operations differ in their handling of null and other key types as compared to
normal array access. Instead of always unioning null to the result (if PHP
array) in array_like_elem, allow the "default" value to be specified, which
gives us more flexibility.

Reviewed By: markw65

Differential Revision: D8966078

fbshipit-source-id: 30b4a6cfaf0834415729aa2fb8b1b678aec360b7