Enable integer shape key declaration (typechecker-only ) and shape access by integer key
Summary:
Enabling integer shape keys and shape access by integer key for use in typechecking for a new, well-typed regex library in HSL. In this library, Regex patterns will be typed with a parameter that represents the resultant shape of a call to `Regex\match`. Because matches can have both positional and named capture groups, this shape will have both integer and string keys.

Integer shape keys will *not* be enabled in WWW. They will still trigger a parser error.

**summary of code changes**

- Enabling integer shape keys: Added a `SFlit_int` node to type `shape_field_name` and changed `SFlit`, originally meant for string literals, to `SFlit_str`. Changed typechecker and bytecode emission code to handle these appropriately.
- Enabling shape access by integer key: Changed the shape field getter called by array gets on shapes to accept integer keys in typing_utils.ml

Reviewed By: kmeht

Differential Revision: D8899906

fbshipit-source-id: 3f892137780433d6844e4534b546b05d5af03e28