Fix a bug with extracting a type from an awaitable under void_is_type_of_null flag
Summary: We allow an expression of type `?Awaitable<T>` to appear as an argument to `await`, and the result of awaiting has the type `?T`.  This logic wasn't replicated correctly under the `void_is_type_of_null` flag: if we await `Awaitable<T> | void` we would produce the result of type `T | _`, which is equivalent to `T`.  Basically, I replaced `void` with `?_` but forgot to wrap the result of extracting the type from `?_` (i.e., `_`) into an option.

Reviewed By: kmeht

Differential Revision: D9013354

fbshipit-source-id: 8addaf012d0326630125c5700b35cf68a6c9d844