Handle methods that return `void` or `Awaitable<void>`
Summary:
When the method returns `void` or `Awaitable<void`, the deprecated wrapper omits `return` in its call to the newly named function.

NOTE: This doesn't take into account type aliases. So if a non-async function has a return type `foo` that happens to be `Awaitable<void>`, the user will still have to manually remove `return` from the generated deprecated wrapper.

Reviewed By: arxanas

Differential Revision: D8986240

fbshipit-source-id: deaab39fd98af4e85a855d16b14d24e495e76e5e