SignatureSearchService
Summary:
Implementation of Signature Search for the Hack standard library and appendum to hh-search.

This is done after the Type-Decl phase in both a Full init and a Saved state init.

This is an extension to the Symbol Search service that will search across the Hack Standard Library for function signatures matching a user's query.

Query format:
function(parameter_type, ...):return_type
Users can also use wildcards in the form of an underscore. "_".
Ex. function(int, _):string

Index is built using a mapping of "search terms" to function names.
Each type within a user query or a function definition is transformed into "search terms" (characteristics of the function).

Ex. function(int, int):void
 becomes { "arity=2"; "arg1<:int"; "arg2<:int"; "ret<:void" }

Differential Revision: D8731084

fbshipit-source-id: 67e2b34421d7920c51c9ba67bbcaaf5dd2a067d2