SignatureSearchIterator
Summary:
Smart Iterator for SignatureSearchIndex

Smart Iterators iterate over files of document IDs until a given constraint is satisfied

In our implementation,
Smart Iterators are constrained based on minimum values, only iterating if the current value is the minimum of one/two values.

Alongside a classic “next”and “peek” function, they implement a “next-gte int ” and “peek-gte int” that continues iteration of a file until end of file or a value is equal or greater to the given integer.

  Ex. And_Iter (file1 [9\n] * file2 [1\n 3\n 9\n])
  And_Iter order of results = 9
  Peek File1: 9, Peek File2: 1
  Next-gte 9 of File2 → next, next, next → 9
  Return 9

  peek of file1 returns 9, peek of file 2 returns 1. We iterate file2 with next-gte 9.
  This iterates twice and returns 9. We are now at end of file for the next call.

Reviewed By: pittsw

Differential Revision: D8883093

fbshipit-source-id: af859dfcf53a86e972e144dc297572ee5093f950