Don't use legacy parser in the vsdebug extension.
Summary:
Nothing was passing the optional releaseUnit param to compile_string, so drop it.
Add a forDebuggerEval param to compile_string and plumb it through to hackc; in hackc, make it check if the unit's pseudomain contained exactly one expression statement, and if so change it to a return statement by munging the bytecode after emission.
Stop using the legacy parser in the vsdebug extension to try to determine if it's safe to prepend a return to the user's input before eval-ing it; pass forDebuggerEval=true to compile_string instead.

Reviewed By: ebluestein

Differential Revision: D9039152

fbshipit-source-id: 415240cbafee0c1e9c70b9be39300f23df6e3ee9