Add RuntimeOption::EvalCheckPropTypeHints
Summary:
Add a new runtime option called EvalCheckPropTypeHints. This option
isn't yet used, this diff just adds the machinery to support it. A setting of 0
means to not enforce property type-hints (current behavior). A setting of 1
means to warn on property type-hint violations. A setting of 2 or 3 means to
raise an error on property type-hint violations. If set to 3, the error cannot
be returned from (meaning enforcement is hard).

Reviewed By: mxw

Differential Revision: D8773245

fbshipit-source-id: a099639fdf433875d2afb96ac1dab8ed14c74c52