Add MOpMode to BaseSC, BaseSL, BaseC, and BaseR
Summary:
Unlike other minstr operations, BaseSC, BaseSL, BaseC, and BaseR didn't
take the MOpMode as an immediate (presumably because they didn't need it). For
the prop type-hint enforcement it will be needed (because we need to do extra
work on base instructions for Define operations), so add it to them.

Reviewed By: swtaarrs

Differential Revision: D8789330

fbshipit-source-id: b6f28f8be89494a2b0b43ae4e176475b968c546f