Add support for property type-hints to TypeConstraint
Summary:
Add support for property type-hints to TypeConstraint along with some other
cleanup. The various TypeConstraint check functions took the current func as a
parameter, but all they did with it was extract the func's class (if
any). Change them to take a Class* directly. Once this was done, a lot of the
helper functions in TypeConstraint became trivial, so remove them. I wanted to
re-use the check logic in a number of different ways, so templatize the check
function depending on how its being used. The "mode" parameter determines if the
check is exact or conservative, if its allowed to consult the enviromment
(loaded classes and type-aliases), if it can autoload, and if the check is for a
property. Also factor out the d/varray checking logic for check failures into a
helper function.

Also cleanup some of the Hack array compat notice functions in runtime-error.cpp
and rename them to make them clear they're Hack array compat functions, not
"Hack array".

Add assertions to the Exception classes that they have the expected type-hints
so that we don't need to worry about doing type-checks on them when setting
properties. If someone changes their definitions, we want things to break so we
know.

Reviewed By: swtaarrs

Differential Revision: D8794224

fbshipit-source-id: 760b03f5ec46ea155b1a17ca1336619a5704d0ee