Implement HHIR ops to perform property type-hint checks
Summary:
Add a new set of HHIR ops to perform property type-hint checks. This
includes the typical set representing known property type-hint failures
(VerifyPropFail and VerifyPropFailHard) as well as one to verify against a Class
(VerifyPropCls). There's also an op to raise a HackArrCompatNotice for d/varray
mismatches on properties (RaiseHackArrPropNotice). Unlike the other cases,
there's also VerifyProp, which handles an entire type-hint check. This is useful
because there's places where we want to emit code to perform a property
type-hint check, but unlike in the param or return value cases, its awkward to
possibly punt at that point.

The existing logic to emit type-hint checks was already complicated because it
dealt with both parameters, in-out, and return value checks. Refactor the
function and templatize it. The logic specific for each type of type-hint is now
provided by various lambdas.

Reviewed By: swtaarrs, markw65

Differential Revision: D8798677

fbshipit-source-id: 1394cfa252fa47f164861ca4bdb690292c06994d