Allow for optional values in PureStore memory effects
Summary:
Sometimes in HHIR you'll have a store which you want store-elim to potentially
remove, but it cannot be modeled as a "pure store" of a single SSATmp. An
example would be a HHIR op which takes two operands, combines them in a single
value, and stores that. Until now, the PureStore memory effects required a
non-null pointer to a SSATmp. Lift this restriction and allow for null pointers
as a value in PureStore. Obviously such stores cannot participate in load-elim,
but they can still be removed if they are proven to never be read from
again. Fix up the few places where we were assuming the value wasn't a null
pointer.

Reviewed By: swtaarrs

Differential Revision: D8813599

fbshipit-source-id: 63e7540135e9a8c10ff68c9c34c3024bc29c964e