Add MInstrPropState to catch prop promotions
Summary:
In order to enforce property type-hints, we need to catch promotions to arrays
or stdclass which would violate the type-hints. The problem is, when the
promotion happens, we no longer know where the base came from and therefore
cannot do the type-check there. In order to handle this, add new state to
MInstrState called MInstrPropState. This contains context about whether the
current base came from a property. It contains the Class*, the properties' slot,
and whether its a static property or not (all the information needed to perform
the type-hint check). In LOWPTR builds it fits into 64-bits.

The state only needs to be looked up when we're doing the promotion, which is
already considered a slow path. However, it needs to be set accurately. We go to
some effort in the JIT to only set it when we might need actually need it. For
example, its only needed for Define member ops and if the base might be one of
the types which can promote. Since its only ever looked at when we're actually
promoting, the state is simply dead if we can prove the promotion cannot happen.

Add a matching alias class representing the MInstrPropState so that store-elim
can also eliminate the stores to it (if we store to the state but never call a
runtime helper, the store can simply be removed). The various HHIR ops which can
result in base promotions have been augmented to take a pointer to a
MInstrPropState. Where necessary, functions which can never promote have been
split out from versions which can (for instance, Define versus not).

Plumbing all this through required a bit more info from the various property
lookup functions (mainly the slot), so add this and cleanup the interfaces a
bit.

Reviewed By: swtaarrs, markw65

Differential Revision: D8814436

fbshipit-source-id: 33d16a3f52b6a8cc752b3bb193c449cad7837e40