Add property type-hint redefinition checking
Summary:
Property redefinitions can cause issues with property type-hint enforcement. If
a class declares a property which redeclares an existing property in the parent,
but the property has a different type-hint, its not clear which type-hint should
be enforced (either choice leads to confusing outcomes). Therefore, when
property type-hint enforcement is enabled, we require that only property
redeclarations with equivalent type-hints are allowed. Two type-hints are
equivalent if they allow the exact same set of values.

When a class is initialized, we check if any of the properties it declares
redeclare an existing property. If so, we check that their two type-hints are
equivalent. If they are not, we either emit a warning or raise an error
(depending on EvalCheckPropTypeHints). Since this is done during class
initialization, it happens at most once per request, per class (though this is
only relevant if we're emitting warnings).

Of course, doing this for every class would be too expensive. Therefore we try
to detect only the cases where a runtime check is necessary. When the class is
loaded and we create the properties, we'll detect if its redeclaring an existing
property, and do a conservative equivalence check on the type-hints. An exact
check is not possible because the equivalence of the two type-hints may differ
from request to request, or may require invoking the autoloader. If the two
type-hints are always equivalent, there's no need to perform a check at
runtime. In addition, HHBBC will annotate properties with AttrNoBadOverride if
it can prove that the two-hints are equivalent (it can do a better job because
of global analysis).

The actual redefinition check is done similarily to property initialization. We
keep a bool in normal RDS for each class, indicating whether the class has
performed a redefinition check in the current request. When creating an object
for that class, we'll check the bool and perform the check if necessary. If
we're running in the TC, we'll only emit checks for the exact properties we're
unsure about.

We were growing a large number of various IR opcodes which just checked if a RDS
handle was initialized, so invent two new IR opcodes CheckRDSInitialized and
MarkRDSInitialized which do just that, and use those in place of the specialized
ones.

Reviewed By: swtaarrs

Differential Revision: D8817772

fbshipit-source-id: 1bfa11c22b3a2a281edcf7bbc04de6ce78988672