Check property initial values for type-hint violations
Summary:
Add the logic to check property initial values for type-hint violations. The
initial values of a property can be initialized two ways. Since the initial
values cannot change during a request, it is redundant to check them everytime
an object is created. Instead they are checked once per request when creating
the Class' property initialization vector (static properties are likewise
checked when initialized). If the initial value isn't a scalar, then when it is
initialized inside the 86pinit or 86sinit function, the value will be type
checked just like a set.

Of course, doing this check for every property would be expensive. So, we only
perform the check for initial values which cannot be proven to always pass the
type-hint. When the Class is created, a check is done to see if the initial
value will always pass the type-hint. If it does, the property is marked with
AttrInitialSatisfiesTC. Such properties are spared the per-request check. In
addition, HHBBC can mark properties AttrInitialSatisfiesTC.

Properties which do not have any user provided initial value provide an
additional problem. The emitter just gives such properties an initial value of
null. However, the property's type-hint may not permit a value of null. HHBBC
can now rewrite system provided property initial values to something which
matches the property's type-hint. This is not possible in all cases, however
(for example, an object has no good non-null initial value). By default, the JIT
will assume a property might be KindOfNull in addition to whatever the type-hint
states (it will not, however, allow null to be set to the property). If
AttrNoImplicitNullable is set, however, this assumption is not made, and the
type-hint is authoritative about what values the property can take. HHBBC will
set AttrNoImplicitNullable as necessary when rewriting property type-hints. It
requires global analysis because a property in a derived class may redeclare a
property in the base with an initial value which is null (and therefore that
property can possibly be null).

Reviewed By: swtaarrs

Differential Revision: D8828575

fbshipit-source-id: 7c336fa4d82a53631462a69a2b95c620f45746ed